requirements.txt
alembic==1.0.0
blinker==1.4
certifi==2018.11.29
chardet==3.0.4
click==6.7
decorator==4.3.0
dominate==2.3.4
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
Mako==1.0.7
MarkupSafe==1.0
more-itertools==4.3.0
nbformat==4.4.0
plotly==3.4.2
psycopg2==2.7.6.1
PyJWT==1.7.1
python-dateutil==2.7.3
python-dotenv==0.10.1
python-editor==1.0.3
pytz==2018.7
requests==2.21.0
retrying==1.3.3
six==1.12.0
SQLAlchemy==1.2.11
traitlets==4.3.2
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
